+++ .github/workflows/black.yml
@@ -1,5 +1,5 @@
 name: black-action
-on: [push, pull_request]
+on: [pull_request]
 jobs:
   linter_name:
     name: runner / black formatter
