--- setup.sh
@@ -1,10 +1,0 @@
-#! /bin/bash
-
-pip install virtualenv
-python3 -m venv venv
-source $1
-
-
-pip install -r requirements.txt
-
-"$@"
