+++ tests/test_sensors.py
@@ -1,10 +1,4 @@
 import os
-import sys
-
-
-# sys.path.append('..')
-
-# from dotenv import dotenv_values
 from fastapi import FastAPI
 from fastapi.testclient import TestClient
 from pymongo import MongoClient
@@ -12,18 +6,11 @@ from dotenv import load_dotenv
 
 load_dotenv()
 
-parent = os.path.abspath(".")
-sys.path.append(parent)
-sys.path.append("../server")
 
-
-print(sys.path)
-
-from App.server.routes.sensor import router as sensor_router
-from App.server.routes.garden import router as garden_router
+from app.routes.sensor import router as sensor_router
+from app.routes.garden import router as garden_router
 
 app = FastAPI()
-# config = dotenv_values(".env")
 app.include_router(sensor_router, tags=["sensor"], prefix="/sensor")
 app.include_router(garden_router, tags=["garden"], prefix="/garden")
